refactor: Service에서 validate 하도록 변경
